config tests refactoring
